Add class for rental options
