+++ rental_options.rb
@@ -1,0 +1,37 @@
+require_relative './rental'
+
+class RentalOptions
+  attr_accessor :rentals
+  def initialize(books, people)
+    @rentals = []
+    @books = books
+    @people = people
+  end
+  
+  def create_rental
+    puts 'Select a book from the following list by number'
+    @books.each_with_index { |book, index| puts "#{index}) Title: '#{book.title}', Author: #{book.author}" }
+    book_number = gets.chomp.to_i
+    puts
+
+    puts 'Select a person from the following list by number (not id)'
+    @people.each_with_index do |person, index|
+      puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+    end
+    person_number = gets.chomp.to_i
+
+    print 'Date: '
+    date = gets.chomp
+    @rentals.push(Rental.new(date, @books[book_number], @people[person_number]))
+    puts 'Rental created successfully'
+  end
+
+  def list_all_rentals
+    print 'ID of person: '
+    id = gets.chomp.to_i
+    puts 'Rentals:'
+    @rentals.each do |rental|
+      puts "Date: #{rental.date}, Book \"#{rental.book.title}\" by #{rental.book.author}" if rental.person.id == id
+    end
+  end
+end
